Move detect_structured_data_type to conversion
