+++ f/common_logic/data_conversion.script.lock
@@ -2,6 +2,7 @@
 attrs==24.2.0
 fiona==1.10.1
 numpy==2.1.3
+lxml==5.3.0
 openpyxl==3.1.5
 pandas==2.2.3
 python-dateutil==2.9.0
