--- f/common_logic/data_validation.script.lock
@@ -1,4 +1,0 @@
-# py311
-lxml==5.3.0
-openpyxl==3.1.5
-pandas==2.2.3
