--- f/common_logic/data_validation.script.yaml
@@ -1,11 +1,0 @@
-summary: 'Shared logic for validating data'
-description: ''
-lock: '!inline f/common_logic/data_validation.script.lock'
-has_preprocessor: false
-kind: script
-no_main_func: true
-schema:
-  $schema: 'https://json-schema.org/draft/2020-12/schema'
-  type: object
-  properties: {}
-  required: []
