+++ tox.ini
@@ -47,7 +47,6 @@ commands =
 
 [testenv:common_logic]
 deps =
-    -r{toxinidir}/f/common_logic/data_validation.script.lock
     -r{toxinidir}/f/common_logic/data_conversion.script.lock
     -r{toxinidir}/f/common_logic/db_transformations.script.lock
     -r{toxinidir}/f/common_logic/db_operations.script.lock
